requirements.txt
bottle==0.12.13
CacheControl==0.12.5
cachetools==3.0.0
certifi==2018.10.15
chardet==3.0.4
Click==7.0
cycler==0.10.0
firebase-admin==2.14.0
google-api-core==1.6.0
google-auth==1.6.1
google-cloud-core==0.28.1
google-cloud-firestore==0.30.0
google-cloud-storage==1.13.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.5
graphviz==0.10.1
grpcio==1.17.0
idna==2.7
imgurpython==1.1.7
itsdangerous==1.1.0
Jinja2==2.10
kiwisolver==1.0.1
MarkupSafe==1.1.0
msgpack==0.6.0
numpy==1.15.2
Pillow==5.3.0
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pygraphviz==1.5
python-dateutil==2.7.3
pytz==2018.7
requests==2.20.1
rsa==4.0
singledispatch==3.4.0.3
six==1.11.0
transitions==0.6.9
urllib3==1.24.1
webencodings==0.5.1
Werkzeug==0.14.1
XlsxWriter==0.7.1
